Convert explicitly to double

+++ handle_error.c
@@ -33,7 +33,7 @@ void formula(long long **values, long long *cntvct, size_t num_threads,
     double numer = (double)(val[0]) * veclen + val[1];
     double denum = (double)(cntvct[tid]) / cntfrq;
     printf("numer: %llf; denum: %llf; ", numer, denum);
-    double gflops = val[0]; // numer / (long long)(1e9) / denum;
+    double gflops = (double)val[0]; // numer / (long long)(1e9) / denum;
     printf("cntvct: %10lld; ", cntvct[tid]);
     for (size_t eid = 0; eid < num_events; eid++) {
       printf("v[%d]: %8lld; ", eid, val[eid]);
